[QD-5] Added ESLint configuration
