+++ server/README.md
@@ -1,0 +1,4 @@
+# Quiz Designer Server
+
+To build all dependencies please use `npm install`. <br>
+To run server please use `npm start`.
