--- server/constants/index.js
@@ -1,5 +1,0 @@
-const errors    =   require('./errors');
-
-module.exports = {
-    errors
-};
