+++ server/package.json
@@ -4,10 +4,8 @@
   "description": "quiz & chill",
   "main": "app.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "_moduleAliases": {
-    "src": "./"
+    "test": "eslint --fix ./src/**/*.js ./src/*.js",
+    "start": "nodemon index.js --exec 'npm run test'"
   },
   "repository": {
     "type": "git",
@@ -31,7 +29,12 @@
     "express": "^4.15.4",
     "jsonwebtoken": "^7.4.3",
     "lodash": "^4.17.4",
-    "module-alias": "^2.0.1",
     "mongoose": "^4.11.9"
+  },
+  "devDependencies": {
+    "eslint": "^4.5.0",
+    "eslint-config-airbnb-base": "^11.3.2",
+    "eslint-plugin-import": "^2.7.0",
+    "nodemon": "^1.11.0"
   }
 }
