+++ server/src/constants/index.js
@@ -1,0 +1,5 @@
+const errors = require('./errors');
+
+module.exports = {
+  errors,
+};
