+++ server/src/models/question.js
@@ -1,0 +1,10 @@
+const mongoose = require('mongoose');
+
+const questionSchema = mongoose.Schema({
+  type: String,
+  question: String,
+  answers: [String],
+  correctAnswers: [Number],
+});
+
+module.exports = mongoose.model('question', questionSchema);
